Removed unused files
