+++ Projects/Authentication/Sources/SEBankID/SEBankIDState.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct SEBankIDState {
+    var autoStartToken: String? = nil
+    var liveQrCodeData: String? = nil
+    var liveQrCodeDate: Date? = nil
+    public init() {}
+}
